// web/src/types/Span.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "../../../gateway/serde_json/JsonValue";
import type { SpanKind } from "./SpanKind";
import type { SpanStatus } from "./SpanStatus";

export type Span = { id: string, parent_id: string | null, run_id: string, kind: SpanKind, name: string, status: SpanStatus, started_at: string, ended_at: string | null, attributes: JsonValue, };